requirements.txt
# Core Django
Django==5.2.6
djangorestframework==3.16.1

# PostgreSQL adapter
psycopg2-binary==2.9.10

# ASGI server (for async support)
asgiref==3.9.1

# Environment variables (optional, but recommended)
python-decouple==3.8

# API documentation (optional)
drf-spectacular==0.27.1

# CORS headers (if you need frontend integration)
django-cors-headers==4.3.1
